+++ src/cli/genpass.rs
@@ -1,0 +1,19 @@
+use clap::Parser;
+
+#[derive(Debug, Parser)]
+pub struct GenPassOpts {
+    #[arg(short, long, default_value_t = 16)]
+    pub length: u8,
+
+    #[arg(long, default_value_t = true)]
+    pub uppercase: bool,
+
+    #[arg(long, default_value_t = true)]
+    pub lowercase: bool,
+
+    #[arg(short, long, default_value_t = true)]
+    pub number: bool,
+
+    #[arg(short, long, default_value_t = true)]
+    pub symbol: bool,
+}
