+++ src/lib.rs
@@ -1,5 +1,5 @@
-mod opts;
-mod process;
+pub mod cli;
+pub mod process;
 
-pub use opts::{Opts, SubCommand};
+pub use cli::{Base64SubCommand, Opts, SubCommand};
 pub use process::*;
